BREAKING: merged STO_ProjectOverview_LegendStudioApplicationPlugin_Extension into LegendStudioApplicationPlugin
