+++ .changeset/hot-dolphins-rest.md
@@ -1,0 +1,5 @@
+---
+'@finos/legend-application-studio': major
+---
+
+**BREAKING CHANGE:** Merged `STO_ProjectOverview_LegendStudioApplicationPlugin_Extension` into `LegendStudioApplicationPlugin`
