+++ packages/legend-application-studio/src/stores/LegendStudioApplicationPlugin.ts
@@ -77,6 +77,11 @@ export type TestableMetadataGetter = (
   editorStore: EditorStore,
 ) => TestableMetadata | undefined;
 
+export type TestRunnerTabRenderer = (
+  selectedTab: string,
+  editorStore: EditorStore,
+) => React.ReactNode | undefined;
+
 export abstract class LegendStudioApplicationPlugin extends LegendApplicationPlugin {
   /**
    * This helps to better type-check for this empty abtract type
@@ -119,6 +124,16 @@ export abstract class LegendStudioApplicationPlugin extends LegendApplicationPlu
    * Get the list of extension for testables
    */
   getExtraTestableMetadata?(): TestableMetadataGetter[];
+
+  /**
+   * Get the list of the supported classifers for test runner tabs.
+   */
+  getExtraTestRunnerTabsClassifiers?(): string[];
+
+  /**
+   * Get the list of renderers for the editor for a test runner tab.
+   */
+  getExtraTestRunnerTabsEditorRenderers?(): TestRunnerTabRenderer[];
 }
 
 export type PureGrammarElementLabeler = (
